ASoC: arizona: Don't block input OSR changes
The moon_osr_put() was returning -EINVAL and logging a kernel error if
any write was made to the control for an analogue input, even if the
value wasn't being changed. This leads to spurious kernel warnings when
writing the control with a valid value.

This was intended to prevent the user choosing low-power settings that
would degrade audio performance, but we can leave the user to decide
what trade-off to make between power and acceptable audio performance.

Change-Id: I40b40b0439dc9320f1585da2708ab9c4d317cfe6
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>