ASoC: arizona: Use compr_buf from DSP rather than stream in IRQ
The trouble with using the wm_adsp_compr_buf pointer from the stream in
the IRQ handler is that we only update the trigger variable when the
stream is open. However, many use-cases require triggering before the
stream is opened. This patch updates to using the pointer stored in the
DSP such that we will always process the trigger event.

Change-Id: I949717599e23731b8dbbcc5952a59171504d587b
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>