irqchip: madera: Use dispatch thread for level-triggered IRQ
The codec can generate interrupts while it is in runtime suspend so
pm_runtime_get_sync() must be called before dispatching the interrupt.
The regmap handler doesn't do this, so we cannot use it to directly
dispatch level-triggered interrupts and must use our local dispatcher
thread for both edge- and level-triggered interrupts.

Change-Id: I408867f8c1a3cb39145a2e483d8f8138d5eebb01
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>