mfd: madera: Allow build of MFD with only some CODECs selected
Currently if we attempt to build the MFD driver with only some
of the CODECs selected we may get a undefined reference to the
patch function for the unselected chips. This is not actually
consistent and seems to work on some branches. This depends on if
the compiler decides to eliminate the code following the BUG_ON,
which it apparently will not always do. The BUG_ON is not really
appropriate here anyway as, whilst we will have checked the chip
type in the SPI/I2C registration, that was based off what we
believed the chip would be, this is the first read of the actual
hardware ID as such this would indicate the wrong chip connected
which should really just fail probe rather than indicate a kernel
bug.

Fix all this by just wrapping the blocks in an if (IS_ENABLED(...
block to ensure the code gets eliminated and return errors if the
chip isn't built in.

Change-Id: I7713ebc31bd45a0891f59aca36719e0415109677
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>