irqchip: madera: Tidy up system suspend/resume handling
We don't need to use the irq_sem variable anymore, when this was
initially introduced the suspend handler that re-enabled the
interrupts was using suspend_late whereas the corresponding resume
function was resume_noirq which caused issues if suspend failed after
our suspend_late had run but before the noirq phase was reached. As we
now are balanced and use noirq on both sides this variable should no
longer be necessary.

Change-Id: I6e8b9d03400256a35040551a7ae9bc26c6682fa3
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>