mfd: arizona: Add sensible return value to some error paths
There are some cases in arizona_dev_init, such as where we don't
recognise the chip ID, in which we head to the error path without
setting a sensible error code in ret. This would lead to the chip
silently failing probe, as it would still return 0. Fix this up by
adding appropriate sets of the return value.

Whilst adding these update the existing paths that do return an error
when the chip is not recognised to use ENODEV, which seems like a better
fit.

Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit 3cddbd44c5a1ecf822c1a286b6a803bc6f564a61)

Change-Id: I4f1de75a3224506990a84fb0819f75d0404ed389
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>