regulator: madera-micsupp: Default regulator to regulating mode
Currently we default the regulator to bypassed mode, however, if
the constraints do not allow bypass then no calls are ever made to
set_bypass. This results in the regulator permanently remaining in
bypass mode, which is very likely not the desired behaviour. Change
the default to regulated mode, such that bypass is only entered when
the constraints allow it.

Change-Id: I3aea529eb673cb54c85672d2f2729590c07dba70
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>