mfd: tacna: Add common support for Cirrus Logic Tacna codecs
This adds the generic core support for Cirrus Logic "Tacna" class codecs.
These are complex audio codec SoCs with a variety of digital and analogue
I/O, onboard audio processing and DSPs, and other features.

These codecs are all based off a common set of hardware IP so can be
supported by a core of common code (with a few minor device-to-device
variations).

Change-Id: I1153216d5231685370d26d31d84023a240828690
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Piotr Stankiewicz <piotrs@opensource.wolfsonmicro.com>