mfd: arizona: Correct probe order for MICVDD
As the MFD code uses the order of the cells during unbind we need to
make sure supplies are last in the list such that all the other children
that use them will have cleaned up by the time the supplies themselves
are destroyed. This does slightly increase probe deferral on boot, but
allows the driver to be unbound cleanly.

Change-Id: I26f8482e2e145e72e4e9398bbb68b1438ad611a7
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>