mfd: madera: Disable IRQs during driver remove
As DCVDD will often be supplied by a child node of the MFD, we
can't call mfd_remove_devices as the first step in arizona_dev_exit
as might be expected (tidy up the children before we tidy up the
MFD). We need to disable and put the DCVDD regulator before we call
mfd_remove_devices, to prevent PM runtime from turning this back on
we also need to disable the PM runtime before we do this. Finally
the IRQs will be cleaned up when the IRQ chip MFD child is removed,
however this will not happen until the other children have been
removes, as they may have registered IRQs.

This creates a window of time where the interrupts are enabled but
the PM runtime, on which the IRQ handler depends, is not available,
any interrupts in this window will go unhandled and fill the log with
failed to resume device messages. To avoid this we simply disable the
main IRQ at the start of arizona_dev_exit, we don't need to actually
handle any IRQs in this window as we are removing the driver.

Change-Id: I1fb0c0e5c0b8bf2aefb5f2d45cdf771ab5d0c98c
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>