ASoC: wm_adsp: Firmware controls should be added as codec controls
We were adding firmware controls as card controls (using
snd_soc_add_card_controls). The DSP is part of a specific codec so
we should be adding them as codec controls. Adding as codec controls
also means that if the codec has a name_prefix it will be added to
the control name, which won't happen when adding as a card control.

As that was the only use of the card pointer in struct wm_adsp it can
be removed.

Change-Id: I6c44f832ba3ac89cf6696cd10113824ade540709
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>