Revert "regulator: arizona-micsupp: initialise micvdd"
This reverts commit 12ba1b6a20c0bb72829fb31134cfa6dbfc063988.

This patch does not work as it compares the maximum voltage the user has
requested with the minimum voltage that will correctly supply all the
micbias's then forces the regulator to that specific voltage. Clearly the
intention was to compare the minimum the user requested and reduce the
range such that it was all above the voltage the micbias's required.
This patch also causes a lot of confusion as several customers have not
understood why the voltage they request for MICVDD is not being applied.
Since it is causing field issues and doesn't work best just to revert.

Change-Id: I31e2585035fda44ea58766a0666b2900105e6a9a
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>