ASoC: marley: Only set the EDRE manual bits for OUT1
The handling of setting EDRE manual when OUT1 clamps are active was
assuming all the other bits in the register are always 0 so can
safely be overwritten with 0 using a regmap_write(). This isn't
good practice and can lead to hard-to-find bugs later if any of the
other bits are non-zero, so change to using an update_bits() so we
only change the bits we need to.

Change-Id: I04d889781909c445d9b2e062eea0c926eaa50240
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>