ASoC: madera: Fix madera_dai_set_sysclk() aliasing
In madera_dai_set_sysclk() we are only supposed to be adding or
removing routes to ASYNCCLK, depending whether the selected source
clock is on ASYNCCLK or SYSCLK domain. The routes to SYSCLK are
always required.

The old implementation didn't deal with the fact that all three
SYSCLKs should be treated the same, and likewise all the ASYNCCLKs
are the same. This meant it could be fooled into removing the routes
to SYSCLK by switching clk_id from one SYSCLK to another SYSCLK.

The patch fixes the code to properly collapse the clock choice down
to a selection between SYSCLK and ASYNCCLK domains, and ensure that
only the ASYNCCLK routes are changed.

Change-Id: I235c5978cedd0f89afb7459fcedbe27dcd828e61
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>