ASoC: cs35l35: reinit_completion before waiting for it
Tommy will fairly frequently generate additional pdn_done IRQs
(most notably after the initial power up of the device). As such
we can actually complete the completion before we wait for it and
the next wait will complete instantly, causing us to not allow
the device time to fully power down. Add a reinit_completion just
before the wait to ensure that all past complete's are out of the
picture and we are actually just waiting for this iteration of
the device powering down.

Change-Id: I8555d63394add31de3fd67d36384ad0a7f396a09
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>