ASoC: cs35l35: Handle shared reset lines
It is quite common to wire up stereo amps to share a reset line, which
ever amp probes first should toggle the reset. As probe defers etc.
might reorder things we need to make sure both amps know about the reset
line.

Allow the request of the reset line to fail if it returns EBUSY and
print a message to say we are assuming the other amp has control of the
reset.

Change-Id: I486044b4a8d44d1d28c773fdfcbaee96251ac390
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>