ASoC: cs35l35: Move register patch after we have identified the chip
It's better to not write anything to a chip's registers until we have
identified that we are indeed talking to the chip we think we are.
Additionally, this means misconfigured systems get a much nicer error
message indicating that we read the wrong chip ID rather than just
register patch failed.

Change-Id: I433bcd890fe4caa9848cb991b11d5eb85473c089
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>