ASoC: cs35l35: Allow user to configure IMON SCALE
On the chip the IMON signal is a full 24-bits however normally only
some of the bits will be sent over the bus. The chip provides a field
to select which bits of the IMON will be sent back, this is the only
feedback signal that has this feature.

Add an additional entry to the cirrus,imon device tree property to
allow the IMON scale parameter to be passed.

Change-Id: I7e9d64a7e3ce48a64aaf6f86e4023dc1f97bc03f
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>