regulator: arizona-micsupp: Move pdata into a separate structure
In preparation for sharing this driver with Madera, move the pdata
for the micsupp regulator out of struct arizona_pdata into a dedicated
pdata struct for this driver. As a result the code in
arizona_micsupp_of_get_pdata() can be made independent of struct arizona.

This patch also updates the definition of struct arizona_pdata and
the use of this pdata in mach-crag6410-module.c

Change-Id: If8a6de212057813d3d1046de0093a61b67e4b5ac
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>