ASoC: cs35l35: Read actually bst_ipk value from regmap
The default for the bst_ipk field is 0x7 and the devicetree/pdata entry
is not mandatory. So if not specified the pdata field will be 0x0 but
the register value will be 0x7, causing the checks in
cs35l35_boost_inductor to set incorrect values.

Correct this issue by using the actual register value rather than the
pdata to make the check.

Change-Id: Idde51a97bb441392876578349c29339d14fa9d08
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>