ASoC: cs35l35: Always check the value of the boost inductor
Currently probe will fail if the boost inductor value (boost_ind) is
not set in a device tree system. However it can be happily left as
zero if using pdata. It would be better to be consistent across both
means of configuring the driver.

Remove the condition on boost_ind being set before calling
cs35l35_boost_inductor to ensure a fail in all cases when boost_ind is
unset.

Change-Id: I2c00b971f912ceaefbf72be5ccebe1cc9a4ca269
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>