ASoC: cs35l35: Add additional delay for reset
Very fast systems may violate the minimum constraints for time the reset
line needs to remain low, or communicate with the device too soon after
releasing the reset. Fix this by adding some delays in to allow the chip
to properly reset, also factor out the reset into a function as it is
likely it will be re-used in later additions to the driver.

Change-Id: Ie98360e8ec26772844a4a7877694fb52f71075a0
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>