ASoC: wm_adsp: Add HALO MPU fault logging
A Halo Core DSP has a memory protection unit that can trap and signal
memory access faults. This patch adds a function that dumps the fault
information.

The interrupt reaches the host via the parent codec interrupt controller
so this fault function is exported to be called by the codec driver.

Change-Id: I46d642b67237a07ae2c1e4d1ded17ad8bfcc7629
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>