ASoC: cs47l94: Toggle VU bits when enabling OUT1 or OUT2
An additional write of VU=1 must be done to the output to ensure
that the mute and volume settings get updated. This must be done when
the codec is powered-up to ensure it is a separate write from any other
writes to the same register, so do it in the output POST_PMU handler.

Change-Id: Ifb3d79cd142d29685595996ff5a8a3fb7b78f016
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>