mfd: tacna: Use separate regmaps for each DSP
The entire register map of the codec is huge and this makes a
single cached regmap slow, especially for debugfs.

This patch separates out the DSP address ranges into their own
regmaps, one per codec. As we only deal in ranges of registers
for the DSPs these new regmaps use range tables instead of
callback functions for the readable/precious indication.

Because all the DSP registers are volatile and we don't need to
apply any patches we can set these regmaps uncached, which makes
access faster. An uncached regmap doesn't need a list of
volatile registers because all accesses are effectively "cache
bypass".

As we now have multiple register maps to setup it is more
convenient to export a function from the codec tables file to
create all register maps instead of doing it all from tacna-spi.c.

The limit of the second set of system control registers has
been brought in to bc7008 so we don't stray into the unused
address range that causes bus faults.

This patch also updates the cs47l94.c codec driver.

Change-Id: Ib184b7d97df75810d63e7a2eaa2b92b6faa5c015
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>