ASoC: wm_adsp: Support restarting a preloaded HALO firmware
Move the HALO MPU and stream arbiter setup from wm_halo_boot_work
to wm_halo_event so that it will be run to initialize the core
each time the DSP is started instead of only when the firmware
is initially loaded. This change makes it possible to stop and
reboot the DSP without reloading the firmware.

Change-Id: I5410042467f9f2641fed397f70438f041a49057c
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>