mfd: tacna: Remove the reset when VDD_D didn't power off
We were forcing a codec reset if VDD_D didn't power off
to ensure all registers are at defaults before the cache
sync. Regmap skips cache values that are at register
default because the chip was supposed to have turned off
so all registers would be at default.

But this reset can cause problems. The only interrupts
that can happen when VDD_D is off are accessory detect,
and because of that we don't need to take any action to
disable other interrupts. But if VDD_D doesn't turn off
it's possible that other interrupts could assert. This
would cause a runtime resume which resets the codec,
losing the interrupt state, and so there aren't any
pending interrupts to handle and it will be treated as
spurious. In the worst case it could be possible to get
into an interrupt storm because we never handle the
cause of the interrupt. The reset also loses information
that would be useful to debug why there was an interrupt.

The codec is always supposed to be used with a dedicated
regulator for VDD_D so that it is guaranteed to turn off
and so there shouldn't be any need to workaround the
problem of VDD_D not turning off.

Change-Id: I020fe87b3ac37cc1e460733d10d01ba43521a6ce
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>