ASoC: cs47l96: Write IN_VU bit after volume or mute changes
A 1 has to be written to IN_VU after any changes to the mute
or volume.

The register containing IN_VU is write-only and only ever needs
a 1 written to it. It is marked volatile so that it will be
uncached, allowing repeated writes of the same value.

For efficiency the input event widget saves the IN_VU until
after all the inputs being activated have been unmuted so
only a single write is needed. For disables the IN_VU is
applied immediately because we want to mute before disabling
to prevent pops.

For the volume controls the write to IN_VU is allowed to fail
if the codec is powered down, since it's irrelevant until the
input is enabled and the power-up of the widget will write
to IN_VU.

Change-Id: I339a984c09d33e1fa53fc7345a0f1d66431d2f0e
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>