ASoC: wm_adsp: Don't init cache from DSP mem if write-only
For controls marked write-only don't initialize the cache from the
content of the DSP memory. We stil need the cache for any new data
that is written to this control, and we need to return something
for a read of the ALSA control because most user-side code assumes
all ALSA controls are readable. The cache is already created zero-
filled so the only change needed is to skip populating it from
DSP memory if the control isn't readable.

Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 04ff40a983e864b586f189b4c3503b6f61263643)

Change-Id: Idecd255ac3d1e0cda3b669ecbb63e58ec549f5e6
Signed-off-by: David Rhodes <david.rhodes@cirrus.com>