mfd: cs47l94: EXT_N memory bits should default to 0
The EXT_N bits should be patched to 0 to save power.
As they are only ever written with the codec powered-up
and don't need any read-modify-write they can be marked
volatile so they won't be cached. This prevents any danger
of a cache sync accidentally powering up in an invalid
order.

Change-Id: I2cf3f9c77318ce938c05984873ab645beb829319
Signed-off-by: Chris Dawson <cdawson@opensource.cirrus.com>
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>