ASoC: wm_adsp: Fix some signedness errors in register access
base and sysclk_reg should be unsigned int, as that is what
regmap takes as a register address.

sysclk_mask is used to mask unsigned register values so
should be unsigned.

sysclk_shift is a shift value so is not allowed to be negative.

Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit e21a5acfab2f13eeffe50d5d5f87c099540a3bf3)

Change-Id: Id5d16518792414c3c6140418cdf37fa381fb11a3
Signed-off-by: David Rhodes <david.rhodes@cirrus.com>