ASoC: wm_adsp: For V3 F/W omit memory region from control names
V3 wmfw files have control names which uniquely identify the control
so there is no need for the memory region to be included in the
ALSA control name. The benefit is that the control can be moved to
a different memory region between firmware releases without affecting
the ALSA control names.

Although V2 firmwares also have unique names the existing name format
is now widely used so we don't want to change it.

Change-Id: I106e2806fdab06e274c0147f1bef30bc985d5047
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>
Signed-off-by: David Rhodes <david.rhodes@cirrus.com>