mfd: tacna: Clear boot SYSCLK error on cs47l96/97
A reset or power-up will cause the SYSCLK err flag to be set.
Ack the flag after boot so that we don't get spurious SYSCLK
error warnings. There is still a pending BOOT_DONE interrupt
to be handled so this won't cause the kernel to see it as an
unhandled interrupt.

Change-Id: I898e2d68b6863da4d7d7b0b719432dafc7799409
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>