ASoC: cs47l96: Add OUT5 source select control
Add control and widget for selecting OUT5 source to be either
the OUT5 signal path or OUT1 signal path.

Note that the OUT1 path has to be powered-up to direct it to
OUT5 but there isn't a separate enable bit control for the OUT1
DAC. So there isn't any need to split the OUT1 PGA into separate
PGA and DAC widgets.

Change-Id: Ib08c56800463b5e2850cbba1e9aee511881dccac
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>