mfd: tacna: Explicitly request VDD_D=1.2v
CS47L96/97 can run at VDD_D=1.0v so would have a regulator
that can be dropped to a lower voltage. But it  must be
booted at 1.2v. Explicitly request 1.2v instead of assuming
that the regulator defaults to this.

Since all other codecs also require 1.2v the request is
unconditional so they all benefit from explicitly ensuring
that the regulator voltage is correct.

Change-Id: I47b0b9b3519689904c6136eeff5d5b9172ed96cf
Signed-off-by: Richard Fitzgerald <rf@opensource.cirrus.com>