ASoC: wm_adsp: Correct and simplify handling of alg offset and length
Currently on Halo based devices we read twice as many registers as
required for the algorithm header blocks. Additionally, the current
code is very unclear as to what units everything is in. Simplify
things and correct the error, by always passing DSP words for both the
offset and the length and then using appropriate helper functions to
convert this to registers.

Change-Id: Ic1d4afb9c96c006c062583bc77171a066a1506d8
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>