mfd: madera: Get all dependencies before registering children
If we EPROBE_DEFER after registering any child devices the children
will be removed and this causes the kernel to retry deferred drivers.
This can cause a loop where every defer immediately causes another
probe.

To avoid this, get the reset gpio and regulators early so that if
they aren't available we can defer before we have created any child
drivers.

Change-Id: Ia734d7ccb9586ae5f7d8371ccda96ff1fa388105
Signed-off-by: Jaswinder Jassal <jjassal@opensource.cirrus.com>