ASoC: madera: Update ASRC to lock out both rates whilst active
The current code will lock out the ASRCx Rate 1 control whilst IN1 on
the ASRC is in use and ASRCx Rate 2 whilst IN2 is in use. However the
ASRC always converts between rates 1 and 2, as such changing either
should be locked out whilst the ASRC is in use.

Since they are always together there is also no need to maintain
separate ASRCxR1CLK and ASRCxR2CLK widgets so combine those as well.

Change-Id: I91c85b879e3f016e97f02446445367712aa98a40
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>