ASoC: madera: Update ISRC to lock out both rates whilst active
The current code will lock out the ISRCx FSL control whilst the DEC
path is in use and ISRCx FSH whilst the INT path is in use. However
the ISRC always converts between FSL and FSH, as such changing
either should be locked out whilst the ISRC is in use.

Since they are always together there is also no need to maintain
separate ISRCxDECCLK and ISRCxINTCLK widgets so combine those
as well.

Change-Id: I055ab68f0d437fb62a29bb6262a7bd99400e3b90
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>