ASoC: wm_adsp: Add support for FW Event controls
FW Event controls enable generic events to be signalled from the firmware
to the driver.  When an assigned IRQ is triggered, the driver will find
all controls of this type for each DSP and then check if they contain
an event.  This event can then be serviced using custom callback function
fwevent_cb.

Change-Id: I8b7384411caa79e3043a820cb3f1ca69e0fe1f0b
Signed-off-by: Stuart Henderson <stuarth@opensource.wolfsonmicro.com>
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Signed-off-by: apandey <ajit.pandey@cirrus.com>