ASoC: madera: Move division of fout to correct check
Currently, we check fout, then divide it by 2, and then save fout. This
causes the check on fout to always fail, although no ill affects result
from this, it does mean we do unnecessary register writes if the FLL is
re-enabled with the same settings. Correct this by moving the division
of fout to be before the check.

Change-Id: I3bcfbc5b7b6f1e32a3894965f25e8d77e8b2c92d
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>