ASoC: wm_adsp: Refactor Halo MPU fault handling
Tenerary operators tend to be frowned on upstream, refactor to remove
the use of them. This does do a little less decode on the MPU fault, but
the majority of the time we will just be passing the values onto the
firmware/hardware team who will probably prefer register values anyway.

Change-Id: I358aeb9093c5513684b6eb092ec20ba275198512
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>