misc: cs40l2x: Disable irq through runtime suspend
GPIO edges prompt the device to render playback and assert an interrupt,
even if the device initially lies in hibernation. If this process occurs
while the system is suspended, the interrupt handler may attempt to read
registers before the bus controller has resumed.

This patch disables and re-enables the irq at runtime suspend and resume,
respectively, so that interrupt processing is deferred until the rest of
the system has resumed.

Change-Id: I345860e1a96c35e39955d3b8f0df632ea75ab235
Signed-off-by: Jeff LaBundy <jeff.labundy@cirrus.com>