misc: cs40l2x: Add support for DSP-controlled amplifier output state
Prior to this patch, specifying cirrus,amp-gnd-stby in device tree
prompted the driver to clamp the amplifier outputs to ground based
on playback state. Doing so sacrificed latency and GPIO support.

This patch diverts amplifier output control to the DSP for firmware
revisions that support this feature. In this case, the DSP controls
amplifier output state automatically without sacrificing latency or
GPIO support.

Change-Id: I034e5d16b31615c8ac6cfbca59de1cc090a0babd
Signed-off-by: Jeff LaBundy <jeff.labundy@cirrus.com>