mfd: clsic: refactor firmware update initiation
There are some cosmetic changes where the code that triggers the
firmware update process and the points where it completes have been
extracted into functions.

The key change is the removal of the forced suspend and pm_get/pm_put
dance, this has been replaced by an explicit call to clsic_msgproc_use
(that will perform a pm_runtime_get and may cause the device to
synchronously switch on) that is reversed with a clsic_msgproc_release
when the flashing operation concludes.

This sequence should reliably initiate the firmware update soft reset
and reenumerate the device.

Change-Id: I39342270f5ad1b28654ec33726b5066cf94f4572
Signed-off-by: Simon Trimmer <simont@opensource.cirrus.com>