misc: cs40l2x: Move teardown to cs40l2x_vibe_mode_worker
If a GPIO trigger interrupts an I2C trigger defined in the wavetable
as a "click," the click is aborted, but the suspend blocker is never
released because playback is still in progress at the time the DSP's
playback state is evaluated.

To solve this problem, move that check from cs40l2x_vibe_pbq_worker
(which is queued only after completion of I2C-triggered playback) to
cs40l2x_vibe_mode_worker (which is queued after any playback event).

Change-Id: I6be405c29286e1049348dd5410ac0c8838839d26
Signed-off-by: Jeff LaBundy <jeff.labundy@cirrus.com>