mfd: madera: Improve handling of Madera unbind
The current unbinding process for Madera has some issues. The trouble
is runtime PM is disabled as the first step of the process, but
some of the drivers release IRQs causing regmap IRQ to issue a
runtime get which fails. To allow runtime PM to remain enabled during
mfd_remove_devices, the DCVDD regulator must remain available. In
the case of external DCVDD's this is simple, the regulator can simply
be disabled/put after the call to mfd_remove_devices. However, in
the case of an internally supplied DCVDD the regulator needs to be
released after all the MFD children except for the regulator child
itself have been removed. This is achieved by having the regulator
driver itself do the disable/put, in this case.

Change-Id: If1df595d8b857662a991c61d915bf78b98fb6654
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>