ASoC: clsic-alg: Fix memory duplication of platform data
Whenever the mfd_add_devices API is called the platform
data provided is duplicated causing mfd children to use
different platform data memory to that of their parents.
This leads to the clsic data structure becoming out of sync.

To avoid this a new API has been added allowing the child
to obtain a pointer to it's associated service handler
struct using the service instance as a unique identifier.

Change-Id: I50a76b4e4c10dfe1bf7853798f8bbf84d746e030
Signed-off-by: Lucas Tanure <tanureal@opensource.cirrus.com>