FROMGIT: opp: Allow lazy-linking of required-opps
The OPP core currently requires the required opp tables to be available
before the dependent OPP table is added, as it needs to create links
from the dependent OPP table to the required ones. This may not be
convenient for all the platforms though, as this requires strict
ordering for probing the drivers.

This patch allows lazy-linking of the required-opps. The OPP tables for
which the required-opp-tables aren't available at the time of their
initialization, are added to a special list of OPP tables:
lazy_opp_tables. Later on, whenever a new OPP table is registered with
the OPP core, we check if it is required by an OPP table in the pending
list; if yes, then we complete the linking then and there.

An OPP table is marked unusable until the time all its required-opp
tables are available. And if lazy-linking fails for an OPP table, the
OPP core disables all of its OPPs to make sure no one can use them.

Tested-by: Hsin-Yi Wang <hsinyi@chromium.org>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>

(cherry picked from commit 7eba0c7641b0009818e469dbfcdd87a0155ab9d4
https://git.kernel.org/pub/scm/linux/kernel/git/vireshk/pm.git opp/linux-next)

BUG=b:129039370
TEST=boot kukui and check cci devfreq governor

Signed-off-by: Hsin-Yi Wang <hsinyi@chromium.org>
Change-Id: I40a60f030dfae434eb791df72ca39971c88efcf2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2635412
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>