mfd: madera: Allow more time for hardware reset
Both manual and power on resets have a brief period where the chip will
not be accessible immediately afterwards. Extend the time allowed for
this from a minimum of 1mS to 2mS based on newer evaluation of the
hardware and ensure this reset happens in all reset conditions. Whilst
making the change also remove the redundant NULL checks in the reset
functions as the GPIO functions already check for this.

Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>
(cherry picked from commit 003035b3eb7085ee7d85f7ed063f10acf66b56d4)

Change-Id: If9c308a05fba34e3d056bb6bff9803fc50b34dc3
Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>