ASoC: cs35l41: Only enable GPIO as SYNC when needed
This is a workaround for high VA current when the
GPIO1 pin is configured as a SYNC pin and the SYNC pin is high
on the passive amplifier.

The workaround is to configure GPIO1 as SYNC immediately
before SYNC is enabled for the passive amplifier, and to
immediately set GPIO1 as a Hi-Z input when SYNC is disabled.

Change-Id: Ie3bb3c1a36700fc7ac7aaa6cba18c45f52c3527c
Signed-off-by: Li Xu <li.xu@cirrus.com>