ASoC: clsic-alg: Update voicectrl codec DAI name
In order to support host buffer control type v1 the stream name
needs to match the buffer name. The compr stream name is set to the
codec_dai name when the stream is opened. The adsp driver creates
the buffer name with dsp hard coded therefore the codec_dai name
has to be dsp rather than vpu.

Change-Id: I27ebbe680ecd6695e88175f130f67de0211a97ff
Signed-off-by: aford <aford@opensource.cirrus.com>