misc: cs40l2x: Multi slot swap open wavetable
This feature is enabled by default and may be disabled using the
cirrus,open-wt-disable devicetree entry.
During initialization, an empty slot is added to the wavetable
for swapping 'virtual' waveforms in and out as needed.  To create
a 'virtual' waveform, the playback queue must be populated first
via cp_trigger_queue, then a special index must be written to
cp_trigger_index (65531).  The data in the queue is saved and added
to a list of 'virtual' waveforms, then the num_waves and
num_virtual_waves sysfs controls are incremented.  The new waveform
may now be played back using cp_trigger_index, gpio1_rise_index or
gpio1_fall_index as if it were a normal waveform in the wavetable.

Change-Id: I41022697b8471ed3b52459c510425fdf4b5d0c2b
Signed-off-by: Tim Howe <tim.howe@cirrus.com>