misc: cs40l2x: Add rewritable waveforms to multi-slot open wavetable
Provides the ability to playback an unlimited number of unique
composite waveforms by overwriting a generic slot in the
wavetable each time a new waveform is sent to the driver.
Usage: Send a string of composite entries to the playback queue
sysfs and then send 65529 to the cp_trigger_index sysfs.

Change-Id: I65baf2920753b377330947d216e684d50f1b0bf8
Signed-off-by: Tim Howe <tim.howe@cirrus.com>