misc: cs40l2x: Add duration support to composite waveforms
Allows the end-user to specify the duration of a waveform within
a composite.  Ex. 3.100.500 in a composite string would indicate
index 3 to be played at 100 amplitude for 500ms.  The ability to
specify a starting delay in milliseconds to a composite has also
been added.  Ex. 0, 250, 3.100.500 would indicate a starting delay
of 250ms.  Optimizations to the open wavetable code were made in
this commit as well in order to allow for the inclusion of these
features.

Change-Id: I1b8db3df79b5bb0a2d6acab0769e60c3eac1df54
Signed-off-by: Tim Howe <tim.howe@cirrus.com>