UPSTREAM: drm/i915: Wait for VLV/CHV/BXT/GLK DSI panel power cycle delay on reboot
As with eDP and LVDS we should also respect the power cycle
delay on DSI panels. We are not using the power sequencer
for these, and we have no optimizations around the sleep
duration, so we just msleep() the whole thing away.

Note that the ICL+ DSI code doesn't seem to have any power
off/power cycle delay handling whatsoever. The only thing it
handles is the power on delay. As that looks pretty busted
in general I won't bother dealing with it for the time being.

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20201001151640.14590-6-ville.syrjala@linux.intel.com
Reviewed-by: Jani Nikula <jani.nikula@intel.com>
(cherry picked from commit f13c2a0032f025fadb57443ed7c2e8eb8e9571df)

BUG=b:179453336
TEST=eDP comes up on brya when built with this commit https://crrev.com/i/3584703

Change-Id: I0a358375de7ba821aef15657433b8a432eac9c87
Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>