CHROMIUM: venus: vdec: workaround for when firmware does not pass crop information
crrev.com/c/2470357 introduced the retrieval of crop information from
the firmware, but as it turns out the firmware does not provide this
information yet for VP8 and VP9 streams. When this happens, fall back to
the previous behavior of setting the crop area to the coded size.

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>

BUG=b:170187774
BUG=b:171148988
TEST=video.DecodeAccelVD.h264_profile_change passes on trogdor.
TEST=video.DecodeAccelVD.vp8 passes on trogdor.
TEST=video.DecodeAccelVD.vp9 passes on trogdor.

Change-Id: Ia71fee5603c78583d62745a4f78d325573949b69
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2486002
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Fritz Koenig <frkoenig@chromium.org>
Commit-Queue: Fritz Koenig <frkoenig@chromium.org>