FROMLIST: firmware: gsmi: Drop the use of dma_pool_* API functions
GSMI driver uses dma_pool_* API functions for buffer allocation
because it requires that the SMI buffers are allocated within 32-bit
physical address space. However, this does not work well with IOMMU
since there is no real device and hence no domain associated with the
device.

Since this is not a real device, it does not require any device
address(IOVA) for the buffer allocations. The only requirement is to
ensure that the physical address allocated to the buffer is within
32-bit physical address space. This is because the buffers have
nothing to do with DMA at all. It is required for communication with
firmware executing in SMI mode which has access only to the bottom
4GiB of memory. Hence, this change switches to using a SLAB cache
created with SLAB_CACHE_DMA32 that guarantees that the allocation
happens from the DMA32 memory zone. All calls to dma_pool_* are
replaced with kmem_cache_*.

In addition to that, all the code for managing the dma_pool for GSMI
platform device is dropped.

Signed-off-by: Furquan Shaikh <furquan@google.com>
Reviewed-by: Ard Biesheuvel <ardb@kernel.org>
(am from https://lore.kernel.org/patchwork/patch/1323449/)
(also found at https://lore.kernel.org/r/20201022071550.1192947-1-furquan@google.com)

BUG=b:170701093
TEST=Verified that gsmi driver is loaded correctly and events can be logged
     for S0ix entry/exit using SMI calls.

Change-Id: I40391afb6e666b18a6169e37739071d76fc31b8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2486920
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Alex Levin <levinale@google.com>
Commit-Queue: Alex Levin <levinale@google.com>
Auto-Submit: Furquan Shaikh <furquan@chromium.org>