UPSTREAM: thunderbolt: Allow disabling XDomain protocol
This allows disabling XDomain protocol completely if the user does not
plan to use the USB4/Thunderbolt peer-to-peer functionality, or for
security reasons.

XDomain protocol is enabled by default but with this commit it is
possible to disable it by passing "xdomain=0" as module parameter (or
through the kernel command line).

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Acked-by: Yehezkel Bernat <YehezkelShB@gmail.com>
(cherry picked from commit 5ca67688256a0aeede5cba288eaef4d8e4a9e622)

BUG=b:186450411
TEST=Build, boot and confirm thunderbolt driver is loaded on brya

Change-Id: I3273e991496e30293b00f9b3b97c496f6063ef3f
Signed-off-by: Azhar Shaikh <azhar.shaikh@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2852778
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>